% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVcluster.R
\name{CNA.out.eachcell}
\alias{CNA.out.eachcell}
\title{CNA output for each cell}
\usage{
CNA.out.eachcell(mean.matrix, log2R.NRC, cluster.index, cutoff = 0.5, L = 100)
}
\arguments{
\item{mean.matrix}{The cluster mean matrix estimated from FLCNA R function.}

\item{log2R.NRC}{Log2R data from normalization of original read counts.}

\item{cluster.index}{Cluster index for all the cells.}

\item{cutoff}{Cutoff value to further control the number of CNAs, the larger value of cutoff, the smaller number of CNAs. The default is 0.35.}

\item{L}{Repeat times in the EM algorithm, defaults to 100.}
}
\value{
The return is the clustered CNA segments by presenting the start position and end position using CNA marker index, and the copy number states.
\item{state}{The CNA states assigned.}
\item{start}{The start point for CNAs.}
\item{end}{The end point for CNAs.}
\item{width}{The width for CNAs.}
\item{sample}{Sample index.}
}
\description{
This function clusters the identified change-points to make final CNA calling for each cell. The potential CNA segments between two neighbor candidate change-points are assigned to different copy number states according to the estimated mean matrix from FLCNA R function and log2R data for each cell. We use three clusters including duplication, normal state and deletion. A Gaussisan Mixture Model based clustering strategy was applied to assign each segment to the most likely cluster/state.
}
