% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVcluster.R
\name{CNA.out}
\alias{CNA.out}
\title{Shared CNA output}
\usage{
CNA.out(mean.matrix, LRR, Clusters, ref, cutoff = 0.4, L = 100)
}
\arguments{
\item{mean.matrix}{The cluster mean matrix estimated from FLCNA R function.}

\item{cutoff}{Cutoff value to further control the number of CNAs, the larger value of cutoff, the smaller number of CNAs. The default is 0.35.}

\item{L}{Repeat times in the EM algorithm, defaults to 100.}
}
\value{
The return is the clustered CNA segments with start position, end position and copy number states.
\item{state}{The CNA states assigned.}
\item{start}{The start point of CNAs.}
\item{end}{The end point of CNAs.}
\item{chr}{Chromosome of CNAs.}
\item{width}{The width of CNAs.}
}
\description{
This function clusters the identified change-points to make final CNA calling. The potential CNA segments between two neighbor candidate change-points are assigned to different copy number states according to the estimated mean matrix from FLCNA R function. We use three clusters including duplication, normal state and deletion. A Gaussisan Mixture Model based clustering strategy was applied to assign each segment to the most likely cluster/state.
}
