% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVcluster.R
\name{CNAcluster}
\alias{CNAcluster}
\title{CNAcluster}
\usage{
CNAcluster(Y, cp, L)
}
\arguments{
\item{Y}{The numeric vector of the intensities of markers, which is the estimated mean vector in our study.}

\item{cp}{The numeric vector of the position index for the identified change-points.}

\item{L}{Repeat times in the EM algorithm, defaults to 100.}
}
\value{
The return is the clustered CNA segments with the start position and end position, length of the CNA and the copy number states (duplication or deletion). It also returns a vector of final candidates of change-points.
\item{newcp}{The final list of change-points.}
\item{h}{The bandwidth used for the identification of change-points.}
\item{CNA.state}{Copy number state for each CNA.}
\item{CNA.start}{Start position of each CNA.}
\item{CNA.end}{End position of each CNA.}
}
\description{
This function clusters CNAs into different states using a Gaussian Mixed Model based clustering strategy.
}
