% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCNA_normalization.R,
%   R/FLCNA_normalization_ref.R
\name{CorrectGC}
\alias{CorrectGC}
\title{GC content correction}
\usage{
CorrectGC(RC, GCContent, step)

CorrectGC(RC, GCContent, step)
}
\arguments{
\item{RC}{A p-dimensional data vector. The read counts for a sample.}

\item{GCContent}{A p-dimensional vector with gc concent.}

\item{step}{Step value, a constant, used in the GC correlation procedure.}
}
\value{
The output for GC content correlation.

The output for GC content correlation.
}
\description{
Normalization according to GC content.

Normalization according to GC content.
}
