% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCNA_normalization.R,
%   R/FLCNA_normalization_ref.R
\name{CorrectMAP}
\alias{CorrectMAP}
\title{Mapp correction}
\usage{
CorrectMAP(RC, MAPContent, step)

CorrectMAP(RC, MAPContent, step)
}
\arguments{
\item{RC}{A P-dimensional data vector. The read counts for a sample.}

\item{MAPContent}{A p-dimensional vector with mappability.}

\item{step}{Step value, a constant, used in the Mapp correlation procedure.}
}
\value{
The output Mapp correlation.

The output Mapp correlation.
}
\description{
Normalization according to mappability.

Normalization according to mappability.
}
