% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCNA_normalization.R,
%   R/FLCNA_normalization_ref.R
\name{CorrectSize}
\alias{CorrectSize}
\title{bin size correction}
\usage{
CorrectSize(RC, L, step)

CorrectSize(RC, L, step)
}
\arguments{
\item{RC}{A P-dimensional data vector. The read counts for a sample.}

\item{L}{The bin size used in the data the default is 100,000.}

\item{step}{Step value, a constant, used in the bin size correlation procedure.}
}
\value{
The output bin size correlation.

The output bin size correlation.
}
\description{
Normalization according to bin size. Since bin size is consistant in all the markers, bin size will not affect normalization results in this study.

Normalization according to bin size. Since bin size is consistant in all the markers, bin size will not affect normalization results in this study.
}
