% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_fn.R
\name{count.mu}
\alias{count.mu}
\title{count.mu}
\usage{
count.mu(mu.j, eps.diff)
}
\arguments{
\item{mu.j}{Mean vector.}

\item{eps.diff}{Lower bound of mean difference.}
}
\description{
Computing the number of unique cluster means for each dimension, which was used in computing BIC or GIC.
}
