% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_fn.R
\name{dmvnorm_log}
\alias{dmvnorm_log}
\title{dmvnorm_log}
\usage{
dmvnorm_log(index, mu, sigma, y)
}
\arguments{
\item{index}{Row index of mu.}

\item{mu}{K by p matrix, each row represents one cluster mean.}

\item{sigma}{p by p covariance matrix (assume same covariance for each cluster).}

\item{y}{n by p data matrix.}
}
\description{
Used in sapply to find all the densities
}
