% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCNA.R
\name{FLCNA}
\alias{FLCNA}
\title{FLCNA analysis}
\usage{
FLCNA(
  tuning = NULL,
  K = NULL,
  lambda = c(5),
  Y,
  N = 100,
  kms.iter = 100,
  kms.nstart = 100,
  adapt.kms = FALSE,
  eps.diff = 1e-05,
  eps.em = 1e-05,
  iter.LQA = 20,
  eps.LQA = 1e-05,
  cutoff = 0.5,
  L = 100,
  model.crit = "bic"
)
}
\arguments{
\item{tuning}{A 2-dimensional vector or a matrix with 2 columns, the first column is the number of clusters \eqn{K} and the second column is the tuning parameter \eqn{\lambda} in the penalty term. If this is missing, then \code{K} and \code{lambda} must be provided.}

\item{K}{The number of clusters \eqn{K}.}

\item{lambda}{The tuning parameter \eqn{\lambda} in the penalty term. The default is 5.}

\item{Y}{A p-dimensional data matrix. Each row is an observation.}

\item{N}{The maximum number of iterations in the EM algorithm. The default value is 100.}

\item{kms.iter}{The maximum number of iterations in kmeans algorithm for generating the starting value for the EM algorithm.}

\item{kms.nstart}{The number of starting values in K-means.}

\item{adapt.kms}{A indicator of using the cluster means estimated by K-means to calculate the adaptive parameters. The default value is FALSE.}

\item{eps.diff}{The lower bound of pairwise difference of two mean values. Any value lower than it is treated as 0.}

\item{eps.em}{The lower bound for the stopping criterion in the EM algorithm.}

\item{iter.LQA}{The number of iterations in the estimation of cluster means by using the local quadratic approximation (LQA).}

\item{eps.LQA}{The lower bound for the stopping criterion in the estimation of cluster means.}

\item{cutoff}{Cutoff value to further control the number of CNAs besed on mean matrix from FL model. Larger cutoff value, less CNAs.}

\item{L}{Repeat times in the EM algorithm while outputing CNA data, defaults to 100.}

\item{model.crit}{The criterion used to select the number of clusters \eqn{K}. It is either `bic' for Bayesian Information Criterion or `gic' for Generalized Information Criterion.}
}
\value{
This function returns the esimated parameters and some statistics of the optimal model within the given \eqn{K} and \eqn{\lambda}, which is selected by BIC when \code{model.crit = 'bic'} or GIC when \code{model.crit = 'gic'}.
\item{K.best}{The optimal number of clusters.}
\item{mu.hat.best}{The estimated cluster means in the optimal model.}
\item{sigma.hat.best}{The estimated covariance in the optimal model.}
\item{alpha.hat.best}{posterior probabilities in the optimal model.}
\item{p.hat.best}{The estimated cluster proportions in the optimal model.}
\item{s.hat.best}{The clustering assignments using the optimal model.}
\item{lambda.best}{The value of tuning hyperparameter lambda that provide the optimal model.}
\item{gic.best}{The GIC of the optimal model.}
\item{bic.best}{The BIC of the optimal model.}
\item{llh.best}{The log-likelihood of the optimal model.}
\item{ct.mu.best}{The degrees of freedom in the cluster means of the optimal model.}
\item{K}{The input k values.}
\item{lambda}{The input lambda values.}
\item{mu.hat}{The estimated cluster means for each parameter combination.}
\item{sigma.hat}{The estimated covariance for each parameter combination.}
\item{p.hat}{The estimated cluster proportions for each parameter combination.}
\item{s.hat = s.hat}{The clustering assignments for each parameter combination.}
\item{gic}{The GIC values for each parameter combination.}
\item{bic}{The BIC values for each parameter combination.}
\item{llh}{The log-likelihood values for each parameter combination.}
\item{ct.mu}{The degrees of freedom in the cluster means for each parameter combination.}
}
\description{
Simultaneous CNA detection and subclone identification using single cell DNA sequencing data.
}
\examples{
Y <- matrix(rnorm(10000, 0, 0.5),10, 1000)
output <- FLCNA(K = c(1:2), lambda = c(2,3), Y=Y)
output

}
