% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCNA_LQA.R
\name{FLCNA_LQA}
\alias{FLCNA_LQA}
\title{Esimation of mean matrix}
\usage{
FLCNA_LQA(
  k,
  K1,
  index.max,
  y,
  mu.t.all,
  mu.no.penal,
  sigma.all,
  alpha,
  lambda,
  iter.num = 20,
  eps.LQA = 1e-05,
  eps.diff = 1e-05
)
}
\arguments{
\item{k}{k-th cluster index.}

\item{K1}{K1 cluster in total.}

\item{index.max}{Initial cluster index assigned accroding to posterior probability.}

\item{y}{n by p data matrix.}

\item{mu.t.all}{K by p mean matrix from previous EM-step.}

\item{mu.no.penal}{K by p mean matrix of unpenalized estimates (lambda=0).}

\item{sigma.all}{p by p diagnal covariance matrix.}

\item{alpha}{n by K posterior probability matrix.}

\item{lambda}{Tuning parameter.}

\item{iter.num}{Max iterations in local quadratic approximation (LQA).}

\item{eps.LQA}{LQA stop criterion.}

\item{eps.diff}{Lower bound of mean difference.}
}
\value{
Estimated mu hat for k-th cluster with totally K1 cluster.
}
\description{
Esimation of mean matrix based on local quadratic approximation (LQA).
}
