% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCNA_normalization.R
\name{FLCNA_normalization_ref}
\alias{FLCNA_normalization_ref}
\title{FLCNA normalization with reference}
\usage{
FLCNA_normalization_ref(Y, bin_size = 1e+05, gc, map, ref_id)
}
\arguments{
\item{Y}{A p-dimensional data matrix. Each row is an observation.}

\item{bin_size}{The bin size used in the data the default is 100,000.}

\item{gc}{A p-dimensional vector with gc concent.}

\item{map}{A p-dimensional vector with mappability.}

\item{ref_id}{cells used as reference.}
}
\value{
The log2Rdata used for main step for the FLCNA method.
}
\description{
Normalization function used in FLCNA.
}
