% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCNA_QC.R
\name{FLCNA_QC}
\alias{FLCNA_QC}
\title{FLCNA_QC}
\usage{
FLCNA_QC(Y_raw, ref_raw, mapp_thresh = 0.9, gc_thresh = c(20, 80))
}
\arguments{
\item{Y_raw}{raw read count matrix.}

\item{ref_raw}{raw GRanges object with corresponding GC content and mappability for quality control.}

\item{mapp_thresh}{scalar variable specifying mappability of each genomic bin. Default is \code{0.9}.}

\item{gc_thresh}{vector specifying the lower and upper bound of GC content threshold for quality control. Default is \code{20-80}.}
}
\value{
A list with components after quality control.
\item{Y}{Read depth matrix after quality control.}
\item{ref}{A GRanges object specifying whole genomic bin positions after quality control.}
}
\description{
Perform QC step on single cells and bins.
}
