% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{flowCalcCpp}
\alias{flowCalcCpp}
\title{Matrix calculation in RcppArmadillo.}
\usage{
flowCalcCpp(Am, Cm)
}
\arguments{
\item{Am}{matrix}

\item{Cm}{matrix}
}
\value{

}
\description{
Matrix calculation in RcppArmadillo.
}
