% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gausianMixture.R
\name{gausianMixture}
\alias{gausianMixture}
\title{Gaussian Mixture Model for CNA clustering}
\usage{
gausianMixture(x, cp, priors, L, st)
}
\arguments{
\item{x}{The vector of the estimated mean of markers.}

\item{cp}{The vector of the marker index of the identified change-points.}

\item{priors}{Given initial parameters for the EM algorithm.}

\item{L}{Repeat times in the EM algorithm. Defaults to 100.}

\item{st}{Number of assumed states in the EM algorithm.}
}
\value{
The return is the clustered CNA segments with the start position and end position using CNA marker index, and the copy number states. It also returns a vector of final candidates of change-points.
\item{p.final}{Probability of falling into each state for each CNA segment after convergence.}
\item{mu.final}{Segment means of each state after convergence.}
\item{cp.final}{List of change-points after EM algorithm.}
\item{index.final}{The index of change-points.}
\item{state.new}{Assigned copy number state for each CNA.}
}
\description{
Gaussian Mixture Model is applied to assign each segment to the most likely cluster/state.
}
