% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVcluster.R
\name{getState}
\alias{getState}
\title{CNA states}
\usage{
getState(EM = EM)
}
\arguments{
\item{EM}{The output of Gaussian Mixture Model for clustering.}
}
\value{
The return is the estimated CNA information.
\item{CNA.state}{Copy number state for each CNA.}
\item{CNA.start}{Start position of each CNA.}
\item{CNA.end}{End position of each CNA.}
}
\description{
This function uses output of Gaussian Mixture Model to obtain different CNA states.
}
