% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gausianMixture.R
\name{updateEM}
\alias{updateEM}
\title{Update parameters using EM algorithm}
\usage{
updateEM(p.in, mu.in, sigma.in, means, sum.x.sq, N, len, st)
}
\arguments{
\item{p.in}{Initial probability for each CNA cluster.}

\item{mu.in}{Initial mean value for each CNA cluster.}

\item{sigma.in}{Initial variance for each CNA cluster.}

\item{means}{Mean value vector for each segment.}

\item{sum.x.sq}{Sum of squared mean values for each segment.}

\item{N}{Number of candiate CNAs.}

\item{len}{Width of candiate CNAs.}

\item{st}{Number of assumed states in the EM algorithm.}
}
\value{
The return is the updated parameters using EM algorithm
\item{p.new}{Updated probability for each CNA cluster.}
\item{mu.new}{Updated mean value for each CNA cluster.}
\item{sigma.new}{Updated variance for each CNA cluster.}
}
\description{
In the Gaussian Mixture Model, parameters will be updated based on EM algorithm.
}
